fix(manager): Maybe fix crash when sending to self
I do not really know why but I was informed that sending a message
crashes the application. When I saw the stack trace, it seemed that the
manager did not have the own contact in the list.
I exchanged the logic to compare the UUIDs instead of checking if the
contact name is empty in the hopes that this would improve the
situation, but I am not really sure that this will fix the crash as I
cannot reproduce it.