fix(freeze): Fix freeze
The problem was the combination of the tokio runtime (used indirectly by
the backend presage) and the glib runtime (used by gtk). This could be
fixed by separating the two different runtimes. This was not that easy
as `presage::Manager` was not `Send`, so I had to construct the `Manager`
in another thread and exchange messagess between the threads.