fix(contact): Fix contacts with special characters
Contacts with special characters had some "isolate"-characers from
Signal arround them. This did not play nicely with the labels in
avatars, which were empty. This commit removes those special characters
around the string, the avatars will therefore be able to show the
initial characters for the contact.

This fixes #62.