chore(refactor): Refactor message item rows
This is a refactor of the old way how the message list was populated.
This was pretty inefficient. Now, it is replaced by an intermediate
`ItemRow` widget which (for some reason, even though it does the same)
makes it considerably more efficient.

This is related to <https://gitlab.com/schmiddi-on-mobile/flare/-/issues/86#note_1504777491>.

This should also make #10 somewhat feasible.