fix(link): Fix linking as secondary device
This was broken due to changes from Signal. This was fixed in libsignal-service-rs in <https://github.com/whisperfish/libsignal-service-rs/pull/240> and <https://github.com/whisperfish/libsignal-service-rs/pull/238> and is being fixed in presage in <https://github.com/whisperfish/presage/pull/185>. Note that currently this depends on an not-yet-upstream fork of presage.

Fixes #98.