feat(): Allow for primary device linking
This MR adds the ability to link as a primary device.

Do not use this, unless:

- You know what you are doing.
- You are knowledgeable enough to activate the primary-device button,
  which I deactivated by default (insert evil laughter here).
- You are knowledgeable enough to switch from staging to production
  servers, which I hardcoded by default (insert evil laughter here).
- You are me.

This MR basically refactors the entirety of the setting up of the
manager, which was previously hard-coded to only allow linking of the
device. In the process, I also removed the link window and added a new
window that gives an introduction of Flare, describes the choice between
setting up as primary or secondary device, enables the user to execute
their preferred setup way, and finally introduces known bugs and the
Matrix room before usage.

A few things that are still TODO:

- Better UI for the setup window.
- Let the user choose between production and staging servers.
- Error handling, e.g. for failed parsing of the phone number.