fix(message): Fix some bugs regarding mentions
This fixes two issues regarding mentions. The first one is that mentions
are actually not highlighted anymore in the text. This probably broke
with the introduction of the special label for messages which allows to
have the timestamp the message was sent at in-line with the actual
message. The bug was caused by overwriting the message attributes when
setting the label, instead of merging the attributes caused by mentions
and the attribute for blank space reserved for the message timestamp. To
fix the issue, the label now saves the message attributes being set and
adds those attributes to the attributes of the underlying label when it
is updated.

The second bug was that message attributes in quotes were using the
attributes of the quoting message instead of the quote itself. The fix is
just a simple change in the blueprint file, taking the attributes for the
label from the quoted message instead of the quoting message.