chores(presage)!: Update the presage store to sqlite
This updates the underlying store used by presage from sled to sqlite.
This change is not backwards compatible, as there currently exists no
migration between the two stores.

While the main reason behind the change from sled to sqlite was, at least
in my eyes, the features which could be implemented using a real
database, the upstream store will still not allow for such features as
there is still binary data stored inside the database, instead of
structured data like I wanted to have. I raised the concern upstream, but
it was mentioned that such a change could be done in the future.

Porting Flare to this new store required a few fixes from my side, which
I contributed upstream in the following PR: https://github.com/whisperfish/presage/pull/321.
The change also required some bigger changes from the Flare side, as
sqlx (the crate presage uses to interface with sqlite) requires a tokio
runtime, and could therefore lead to crashes if the store was used from
the glib thread. That is why I wrapped all (at least I hope I did not
miss any) calls to the store and internal manager in a tspawn. Some of
those would maybe not be required, but better be on the save side in case
the functions of presage change to access the store where they previously
did not.

I also introduced a new error in case creation of the folder holding
the database failed. There is also still one potential to panic in case
this path is not UTF-8, but I don't care about that right now.
Furthermore, presage changed the order of messages being returned from
incrementing timestamps to decrementing timestamps, I therefore had to
throw away iterator reversal I did in Flare. Other than those changes,
the rest of this commit pretty much only replaces references from the
sled store to the sqlite store.