fix(profile): Fix profile name not displayed correctly
Fixes #248 and #263. I'm not completely sure what the cause of this issue is, but
presage sometimes overwrites a contact with an empty contact except for
the phone number which is set. This in particular unsets the profile key,
which we use to fetch profile information. For some reason, presage also
stores the profile key in a separate table. We can get our profile key
from there in case it is unset from a contact.